+++ jumi-launcher/src/test/java/fi/jumi/launcher/remote/RemoteSuiteLauncherTest.java
@@ -42,7 +42,7 @@ public class RemoteSuiteLauncherTest {
     @Test
     public void sends_RunTests_command_to_the_daemon_when_it_connects() {
         SuiteConfiguration config = new SuiteConfigurationBuilder()
-                .addToClassPath(Paths.get("dependency.jar"))
+                .addToClasspath(Paths.get("dependency.jar"))
                 .setTestClasses("FooTest")
                 .freeze();